Erase -> pop_back+push_back Add twigs to tree.h protected: Get rid of Ibody, Icell Use .I inside sort Move nth_element to partition and change interface to take vectors by reference

+++ include/types.h
@@ -12,7 +12,7 @@
 #include "vec.h"
 
 typedef long         bigint;                                    // Big integer type
-typedef double       real;                                      // Real number type
+typedef float        real;                                      // Real number type
 typedef vec<3,real>  vect;                                      // 3-D vector type
 typedef vec<10,real> coef;                                      // Multipole coefficient type
 typedef std::vector<bigint>           Bigints;                  // Vector of big integer types
